requirements.txt
django
gunicorn
drf-spectacular
djangorestframework-simplejwt
django-cors-headers
djangorestframework
django-filter
django_seed
psycopg2-binary
